Convert navigation bar change functions to module
